Fixed Bug 2
The loot spawned to late and the dead body blocks projectiles. I fixed it by disabling the collider of the enemy on death and moved the code for droping loot to the moment the enemy dies. It can be found in the RunnerBehaviour and can be tested be killing an enemy ingame.